+++ src/domain/services/tmdb/get-tmdb-tv-series.ts
@@ -7,6 +7,7 @@ type GetTMDBTvSeriesServiceInput = {
   language: Language
   returnSeasons?: boolean
   returnGenres?: boolean
+  returnCountries?: boolean
 }
 
 const ONE_WEEK_IN_SECONDS = 7 * 24 * 60 * 60
@@ -18,6 +19,7 @@ export async function getTMDBTvSeriesService(
     language,
     returnSeasons = false,
     returnGenres = false,
+    returnCountries = false,
   }: GetTMDBTvSeriesServiceInput
 ) {
   const cacheKey = `TV_SHOW:${tmdbId}:${language}`
@@ -31,6 +33,7 @@ export async function getTMDBTvSeriesService(
       backdropPath: data.backdrop_path,
       ...(returnSeasons && { seasons: data.seasons }),
       ...(returnGenres && { genres: data.genres }),
+      ...(returnCountries && { countries: data.production_countries }),
     }
   }
 
@@ -43,5 +46,6 @@ export async function getTMDBTvSeriesService(
     backdropPath: data.backdrop_path,
     ...(returnSeasons && { seasons: data.seasons }),
     ...(returnGenres && { genres: data.genres }),
+    ...(returnCountries && { countries: data.production_countries }),
   }
 }