+++ src/http/controllers/user-stats.ts
@@ -8,6 +8,7 @@ import { getUserMostWatchedSeriesService } from '@/domain/services/user-stats/ge
 import { languageQuerySchema } from '../schemas/common'
 import { getUserWatchedGenresService } from '@/domain/services/user-stats/get-user-watched-genres'
 import { getUserWatchedCastService } from '@/domain/services/user-stats/get-user-watched-cast'
+import { getUserWatchedCountriesService } from '@/domain/services/user-stats/get-user-watched-countries'
 
 export async function getUserStatsController(
   request: FastifyRequest,
@@ -82,6 +83,23 @@ export async function getUserWatchedCastController(
 ) {
   const { id } = getUserDefaultSchema.parse(request.params)
   const result = await getUserWatchedCastService({ userId: id, redis })
+
+  return reply.status(200).send(result)
+}
+
+export async function getUserWatchedCountriesController(
+  request: FastifyRequest,
+  reply: FastifyReply,
+  redis: FastifyRedis
+) {
+  const { id } = getUserDefaultSchema.parse(request.params)
+  const { language } = languageQuerySchema.parse(request.query)
+
+  const result = await getUserWatchedCountriesService({
+    userId: id,
+    redis,
+    language,
+  })
 
   return reply.status(200).send(result)
 }